requirements.txt
future~=0.18.2
Pillow~=7.2.0
numpy~=1.19.2
opencv-python~=4.4.0.44
pip~=20.2.4
Brotli~=1.0.9
Flask~=1.1.2
plotly~=4.10.0
Werkzeug~=1.0.1
dash~=1.16.0
h5py~=3.1.0
pytz~=2020.4
scipy~=1.5.2
matplotlib~=3.3.2
Jinja2~=2.11.2
itsdangerous~=1.1.0
click~=7.1.2
setuptools~=50.3.2
MarkupSafe~=1.1.1
joblib~=0.17.0
pandas~=1.1.2
python-dateutil~=2.8.1
six~=1.15.0
retrying~=1.3.3
scikit-image~=0.17.2
certifi~=2020.11.8
tifffile~=2020.10.1
cycler~=0.10.0
seaborn~=0.11.1
imageio~=2.9.0
networkx~=2.5
sklearn~=0.0
scikit-learn~=0.23.2
threadpoolctl~=2.1.0
decorator~=4.4.2
pyparsing~=2.4.7
kiwisolver~=1.3.1